# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/campaigns/seos/{campaign_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get the list of seo campaign you created",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Master - Campaign SEO"
                ],
                "summary": "List SEO Campaign",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Campaign ID",
                        "name": "campaign_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Page of list data",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Limitting data you want to get",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Find your data with keywords",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "created_at:desc",
                        "description": "Ordering data",
                        "name": "orderBy",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Request success",
                        "schema": {
                            "$ref": "#/definitions/commonschema.ResponseHTTP"
                        }
                    },
                    "400": {
                        "description": "Request failure",
                        "schema": {
                            "$ref": "#/definitions/commonschema.ResponseHTTP"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create new seo campaign data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Master - Campaign SEO"
                ],
                "summary": "Create SEO Campaign",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Campaign ID",
                        "name": "campaign_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "SEO campaign payload",
                        "name": "campaignSeoPayload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/masterschema.CampaignSeoPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Request success",
                        "schema": {
                            "$ref": "#/definitions/commonschema.ResponseHTTP"
                        }
                    },
                    "400": {
                        "description": "Request failure",
                        "schema": {
                            "$ref": "#/definitions/commonschema.ResponseHTTP"
                        }
                    }
                }
            }
        },
        "/api/campaigns/seos/{campaign_id}/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get detail of seo campaign you created",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Master - Campaign SEO"
                ],
                "summary": "Detail SEO Campaign",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Campaign ID",
                        "name": "campaign_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID of your data",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Request success",
                        "schema": {
                            "$ref": "#/definitions/commonschema.ResponseHTTP"
                        }
                    },
                    "400": {
                        "description": "Request failure",
                        "schema": {
                            "$ref": "#/definitions/commonschema.ResponseHTTP"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update existing seo campaign data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Master - Campaign SEO"
                ],
                "summary": "Update SEO Campaign",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Campaign ID",
                        "name": "campaign_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID of your data",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "SEO campaign payload",
                        "name": "campaignSeoPayload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/masterschema.CampaignSeoPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Request success",
                        "schema": {
                            "$ref": "#/definitions/commonschema.ResponseHTTP"
                        }
                    },
                    "400": {
                        "description": "Request failure",
                        "schema": {
                            "$ref": "#/definitions/commonschema.ResponseHTTP"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete existing seo campaign data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Master - Campaign SEO"
                ],
                "summary": "Delete SEO Campaign",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Campaign ID",
                        "name": "campaign_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID of your data",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Request success",
                        "schema": {
                            "$ref": "#/definitions/commonschema.ResponseHTTP"
                        }
                    },
                    "400": {
                        "description": "Request failure",
                        "schema": {
                            "$ref": "#/definitions/commonschema.ResponseHTTP"
                        }
                    }
                }
            }
        },
        "/api/campaigns/{workspace_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get the list of campaigns you created",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Master - Campaigns"
                ],
                "summary": "List Campaigns",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Workspace ID",
                        "name": "workspace_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Page of list data",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Limitting data you want to get",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Find your data with keywords",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "created_at:desc",
                        "description": "Ordering data",
                        "name": "orderBy",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Request success",
                        "schema": {
                            "$ref": "#/definitions/commonschema.ResponseHTTP"
                        }
                    },
                    "400": {
                        "description": "Request failure",
                        "schema": {
                            "$ref": "#/definitions/commonschema.ResponseHTTP"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create new campaign data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Master - Campaigns"
                ],
                "summary": "Create Campaign",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Workspace ID",
                        "name": "workspace_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "campaign payload",
                        "name": "campaignPayload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/masterschema.CampaignPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Request success",
                        "schema": {
                            "$ref": "#/definitions/commonschema.ResponseHTTP"
                        }
                    },
                    "400": {
                        "description": "Request failure",
                        "schema": {
                            "$ref": "#/definitions/commonschema.ResponseHTTP"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "commonschema.ResponseHTTP": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "error": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "masterschema.CampaignFormAttributePayload": {
            "type": "object",
            "required": [
                "label",
                "value"
            ],
            "properties": {
                "id": {
                    "type": "string"
                },
                "is_default": {
                    "type": "boolean",
                    "default": false
                },
                "label": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
        "masterschema.CampaignFormPayload": {
            "type": "object",
            "required": [
                "form_id",
                "title"
            ],
            "properties": {
                "attributes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/masterschema.CampaignFormAttributePayload"
                    }
                },
                "default_value": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "form_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_multiple": {
                    "type": "boolean",
                    "default": false
                },
                "is_required": {
                    "type": "boolean",
                    "default": false
                },
                "placeholder": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "masterschema.CampaignPayload": {
            "type": "object",
            "required": [
                "forms",
                "title"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "forms": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/masterschema.CampaignFormPayload"
                    }
                },
                "is_publish": {
                    "type": "boolean",
                    "default": false
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "masterschema.CampaignSeoPayload": {
            "type": "object",
            "properties": {
                "access_key": {
                    "type": "string"
                },
                "event": {
                    "type": "string"
                },
                "platform": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer \" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}